Abstract navigation data logic into a shared middleware
