+++ lib/routes/url-updater.js
@@ -3,27 +3,12 @@
 const express = require('express');
 const Raven = require('raven');
 const cacheControl = require('@financial-times/origami-service').middleware.cacheControl;
-const navigation = require('../../data/navigation.json');
+const addNavigationToRequest = require('../middleware/add-navigation-to-request');
 const parseModulesParameter = require('../url-updater/parse-modules-parameter');
 const evaluateModules = require('../url-updater/evaluate-modules');
 const parseBuildServiceUrl = require('../url-updater/parse-build-service-url');
 const updateUrlForResults = require('../url-updater/update-url-for-results');
 const UserError = require('../utils/usererror');
-
-function addNavigationToRequest() {
-    return (request, response, next) => {
-        const path = request.path.split('?')[0];
-        navigation.items.map(item => item.current = false);
-        const item = navigation.items.find(item => {
-            return path === '/' + item.href;
-        });
-        if (item) {
-            item.current = true;
-        }
-        request.navigation = navigation;
-        next();
-    };
-}
 
 module.exports = app => {
     app.get('/url-updater/', cacheControl({ maxAge: '7d' }), addNavigationToRequest(), (request, response) => {
@@ -34,7 +19,7 @@ module.exports = app => {
         });
     });
 
-    app.post('/url-updater/', cacheControl({ maxAge: '5m' }), addNavigationToRequest(), express.urlencoded({ extended: false }), express.text(), async (request, response) => {
+    app.post('/url-updater/', cacheControl({ maxAge: 0 }), addNavigationToRequest(), express.urlencoded({ extended: false }), express.text(), async (request, response) => {
         try {
             const url = parseBuildServiceUrl(request.body['build-service-url']);
             const brand = url.searchParams.get('brand');
@@ -48,7 +33,7 @@ module.exports = app => {
             response.render('url-updater', {
                 title: 'Origami Build Service',
                 layoutStyle: '',
-                navigation,
+                navigation: request.navigation,
                 buildServiceUrl: url,
                 updatedBuildServiceUrl,
                 hasOutdatedComponents,
