+++ lib/routes/v2/docs/migration.js
@@ -1,23 +1,14 @@
 'use strict';
 
 const cacheControl = require('@financial-times/origami-service').middleware.cacheControl;
-const navigation = require('../../../../data/navigation.json');
+const addNavigationToRequest = require('../../../middleware/add-navigation-to-request');
 
 module.exports = app => {
-	app.get('/v2/docs/migration', cacheControl({maxAge: '7d'}), (request, response) => {
-		const path = request.path.split('?')[0];
-		navigation.items.map(item => item.current = false);
-		const item = navigation.items.find(item => {
-			return path === '/'+item.href;
-		});
-		if (item) {
-			item.current = true;
-		}
-
+	app.get('/v2/docs/migration', cacheControl({ maxAge: '7d' }), addNavigationToRequest(), (request, response) => {
 		response.render('migration', {
 			title: 'Migration Guide - Origami Build Service',
 			layoutStyle: 'o-layout--docs',
-			navigation
+			navigation: request.navigation
 		});
 	});
 };
