+++ lib/routes/v2/index.js
@@ -1,16 +1,15 @@
 'use strict';
 
 const cacheControl = require('@financial-times/origami-service').middleware.cacheControl;
-const navigation = require('../../../data/navigation.json');
+const addNavigationToRequest = require('../../middleware/add-navigation-to-request');
+
 module.exports = app => {
 	// v2 home page
-	app.get('/v2', cacheControl({maxAge: '7d'}), (request, response) => {
-		navigation.items.map(item => item.current = false);
-		navigation.items[0].current = true;
+	app.get('/v2', cacheControl({ maxAge: '7d' }), addNavigationToRequest(), (request, response) => {
 		response.render('index', {
 			title: 'Origami Build Service',
 			layoutStyle: 'o-layout--landing',
-			navigation
+			navigation: request.navigation
 		});
 	});
 };
